requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
eventlet==0.31.0
pillow==8.3.1
py7zr==0.16.2
rarfile==4.0
zipfile36==0.1.3
gunicorn==20.1.0
requests==2.26.0
python-dotenv==0.19.0  # Added this line
gevent 
gevent-websocket
pyunpack 
patool